ready

+++ Event/Models/Respone/Item/CreateItemModel.cs
@@ -9,6 +9,6 @@ namespace Event.Models.Respone.Item
     {
         public string UserId { get; set; }
         public int CategoryId { get; set; }
-        public int PointValue { get; set; }
+        public double PointValue { get; set; }
     }
 }
